% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Birthday.R
\name{Birthday}
\alias{Birthday}
\title{Birthday function}
\usage{
Birthday(x)
}
\arguments{
\item{x}{numeric vector of the number of people in a group}
}
\value{
numeric probability of at least two people sharing a birthday
}
\description{
Birthday function
}
\examples{
Birthday(20:24)
}
