% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt.R
\docType{data}
\name{ddt}
\alias{ddt}
\title{DDT Concentration data}
\format{
## `ddt`
A data frame with 144 rows and 6 columns:
\describe{
  \item{RIVER}{Name of river where sample was taken}
  \item{MILE}{Distance (in miles) from river mouth}
  \item{SPECIES}{Fish species sampled}
  \item{LENGTH}{Fish length (cm)}
  \item{WEIGHT}{Fish weight (g)}
  \item{DDT}{Measured DDT concentration (ppm)}
  ...
}
}
\source{
Provided in class
}
\usage{
ddt
}
\description{
A dataset containing measurements of DDT concentrations in the
Tennessee River Valley
}
\keyword{datasets}
