% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{FIREDAM fire dataset}
\format{
A data frame with 15 rows and 2 variables:
\describe{
  \item{DISTANCE}{Distance of fire from some reference point}
  \item{DAMAGE}{Damage in monetary units or scale}
}
}
\source{
<URL or reference for the data>
}
\usage{
fire
}
\description{
A dataset containing fire-related data from XYZ source.
}
\examples{
NULL
}
\keyword{datasets}
