% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.R
\name{mycltu}
\alias{mycltu}
\title{Central Limit Theorem Simulation for Uniform Distribution}
\usage{
mycltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{Integer. Sample size for each iteration.}

\item{iter}{Integer. Number of iterations.}

\item{a}{Numeric. Lower bound of the uniform distribution. Default is 0.}

\item{b}{Numeric. Upper bound of the uniform distribution. Default is 10.}
}
\value{
A histogram with overlaid density curves: empricial density, theoretical normal approximation,
and the original uniform distribution
}
\description{
Central Limit Theorem Simulation for Uniform Distribution
}
\examples{
mycltu(n = 20, iter = 100000)
}
