% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myf.R
\name{myf}
\alias{myf}
\title{Visualize repeated sampling from a uniform distribution}
\usage{
myf(n, iter = 10, time = 0.5)
}
\arguments{
\item{n}{integer - number of samples per iteration}

\item{iter}{integer - number of iterations to run}

\item{time}{numeric - pause time between plots in seconds}
}
\value{
no return value - function is called for its side effect: generating barplots
}
\description{
Visualize repeated sampling from a uniform distribution
}
\examples{
myf(n = 1000, iter = 10, time = 0.5)
}
