% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Plot and shade normal curve with area}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean of the normal distribution}

\item{sigma}{standard deviation of the normal distribution}

\item{a}{value up to which area is calculated}
}
\value{
a list with mu, sigma, a, and P(X ≤ a)
}
\description{
Plot and shade normal curve with area
}
\examples{
myncurve(mu = 100, sigma = 15, a = 120)

}
