% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{integer - number of available seats on flight}

\item{gamma}{numeric - probability of overbooking}

\item{p}{numeric - probability that passenger shows up}
}
\value{
a named list containing:
\describe{
  \item{nd}{maximum ticket number using discrete binomial distribution}
  \item{nc}{maximum ticket number using normal approximation}
  \item{N}{number of available seats on flight}
  \item{p}{probability that passenger shows up}
  \item{gamma}{overbooking probability}
}
}
\description{
ntickets
}
\examples{
ntickets(N = 400, gamma = 0.02, p = 0.95)
}
